Query msg
I have requested assistance, but the team doesn't seem to know what it is.  And there is not just one but different persons and each seems to have a different solution and reason why it's not executing successfully. I don't know what to do.  Please help.  This is the msg I keep getting.  
Msg 208 Level  16, State 1, Line 3
Invalid object name 'courseInstructor'